Second iteration of refactoring and implementation of missing stuff
