+++ omc3/model/accelerators/skekb.py
@@ -1,11 +1,7 @@
-import os
-
 from model.accelerators.accelerator import (Accelerator, AcceleratorDefinitionError,
                                             AccExcitationMode)
 from utils import logging_tools
-from parser.entrypoint import EntryPoint, EntryPointParameters, split_arguments
 
-CURRENT_DIR = os.path.dirname(__file__)
 LOGGER = logging_tools.get_logger(__name__)
 
 
@@ -16,42 +12,6 @@ class SKekB(Accelerator):
     """
     NAME = "skekb"
     MACROS_NAME = "skekb"
-
-    @classmethod
-    def init_and_get_unknowns(cls, args=None):
-        """ Initializes but also returns unknowns.
-
-         For the desired philosophy of returning parameters all the time,
-         try to avoid this function, e.g. parse outside parameters first.
-         """
-        opt, rest_args = split_arguments(args, cls.get_instance_parameters())
-        return cls(opt), rest_args
-
-    @classmethod
-    def get_class(cls, *args, **kwargs):
-        """ Returns subclass .
-
-        """
-        parser = EntryPoint(cls.get_class_parameters(), strict=True)
-        opt = parser.parse(*args, **kwargs)
-        return cls._get_class(opt)
-
-    @classmethod
-    def get_class_and_unknown(cls, *args, **kwargs):
-        """ Returns subclass and unkown args .
-
-        For the desired philosophy of returning parameters all the time,
-        try to avoid this function, e.g. parse outside parameters first.
-        """
-        parser = EntryPoint(cls.get_class_parameters(), strict=False)
-        opt, unknown_opt = parser.parse(*args, **kwargs)
-        return cls._get_class(opt), unknown_opt
-
-    @classmethod
-    def _get_class(cls, opt):
-        """ Actual get_class function """
-        new_class = cls
-        return new_class
 
     def verify_object(self):  # TODO: Maybe more checks?
         if self.model_dir is None:  # is the class is used to create full response?
